Mock_SWE_Interviewer/requirements.txt
google-cloud
flask
pywebview
requests
wave
openai
python-dotenv